requirements.txt
pandas
numpy
openpyxl
matplotlib
pymysql
sqlalchemy
pyyaml
requests
argparse
geopy
certifi
